# backend/docs/swagger.yaml
definitions:
...
  handler.APIResponse:
    properties:
      data: {}
      msg:
        example: ok
        type: string
    type: object
...
  handler.stationRequest:
    properties:
      code:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      status:
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
paths:
...
  /b/roles/{role}/permissions:
    get:
      consumes:
      - application/json
      description: 获取指定角色的权限列表
      parameters:
      - description: 角色标识
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取角色权限
      tags:
      - b_role
    put:
      consumes:
      - application/json
      description: 更新指定角色的权限列表，受影响用户需重新登录
      parameters:
      - description: 角色标识
        in: path
        name: role
        required: true
        type: string
      - description: 权限列表
        in: body
        name: request
        required: true
        schema:
          properties:
            permissions:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 更新角色权限
      tags:
      - b_role
  /b/stations:
    get:
      consumes:
      - application/json
      description: 分页获取服务站点列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取服务站点列表
      tags:
      - b_station
    post:
      consumes:
      - application/json
      description: 创建新的服务站点
      parameters:
      - description: 站点信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.stationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 创建服务站点
      tags:
      - b_station
  /b/stations/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的服务站点
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 删除服务站点
      tags:
      - b_station
    get:
      consumes:
      - application/json
      description: 获取指定服务站点的详细信息
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: 站点不存在
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取服务站点详情
      tags:
      - b_station
    put:
      consumes:
      - application/json
      description: 更新指定服务站点的信息
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 站点信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.stationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 更新服务站点
      tags:
      - b_station
...
  /b/users/{id}/identities:
    put:
      consumes:
      - application/json
      description: 授予或撤销用户身份，用户需重新登录
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 身份列表
        in: body
        name: request
        required: true
        schema:
          properties:
            identities:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 更新用户身份
      tags:
      - b_role
...
  /c/stations/match:
    get:
      consumes:
      - application/json
      description: 根据经纬度匹配最近的服务站点
      parameters:
      - description: 纬度
        format: float64
        in: query
        name: lat
        type: number
      - description: 经度
        format: float64
        in: query
        name: lng
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: 匹配成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: 匹配服务站点
      tags:
      - public
...
swagger: "2.0"